pyproject.toml
[project]
name = "rdtbkmkchk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "praw>=7.7.1",
    "python-dotenv>=1.0.1",
]
